Standardized code (#8)
- Removed redundant code from abstract classes, in favour of traits;
- Setting invalid error list on exceptions now causes `InvalidArgumentException`
to be thrown, instead of a validation-specific exception.
