+++ test/functional/Exception/AbstractValidationFailedExceptionTest.php
@@ -51,52 +51,4 @@ class AbstractValidationFailedExceptionTest extends TestCase
         $this->assertInstanceOf(static::TEST_SUBJECT_CLASSNAME, $subject, 'Could not create a valid instance');
         $this->assertInstanceOf('Exception', $subject, 'Subject must be a valid exception');
     }
-
-    /**
-     * Tests that validation errors are set and retrieved correctly.
-     *
-     * @since 0.1
-     */
-    public function testGetSetValidationErrors()
-    {
-        $subject = $this->createInstance();
-        $data = array('banana', 'orange', 'apple');
-        $reflection = $this->reflect($subject);
-
-        $this->assertEquals(array(), $reflection->_getValidationErrors(), 'Validation errors could not be set/retrieved correctly');
-
-        $reflection->_setValidationErrors($data);
-        $this->assertEquals($data, $reflection->_getValidationErrors(), 'Validation errors could not be set/retrieved correctly');
-    }
-
-    /**
-     * Tests that settings invalid validation error list results in an exception.
-     * 
-     * @expectedException \Dhii\Validation\Exception\ValidationExceptionInterface
-     *
-     * @since 0.1
-     */
-    public function testSetValidationErrorsThrows()
-    {
-        $subject = $this->createInstance();
-        $data = 'banana';
-        $reflection = $this->reflect($subject);
-
-        $reflection->_setValidationErrors($data);
-    }
-
-    /**
-     * Tests that the validation subject is set and retrieved correctly.
-     *
-     * @since 0.1
-     */
-    public function testGetSetValidationSubject()
-    {
-        $subject = $this->createInstance();
-        $data = 'banana';
-        $reflection = $this->reflect($subject);
-
-        $reflection->_setValidationSubject($data);
-        $this->assertEquals($data, $reflection->_getValidationSubject(), 'Validation subject could not be set/retrieved correctly');
-    }
 }
