+++ test/stub/AbstractValidationFailedException.php
@@ -14,19 +14,4 @@ use Dhii\Validation\Exception\AbstractValidationFailedException as BaseFailedExc
  */
 abstract class AbstractValidationFailedException extends BaseFailedException implements ValidationFailedExceptionInterface
 {
-    /**
-     * @since [*next-version*]
-     *
-     * @param string    $message
-     * @param int       $code
-     * @param Exception $previous
-     * @param mixed     $subject  The validation subject
-     * @param array     $errors   The error list.
-     */
-    public function __construct($message = '', $code = 0, Exception $previous = null, $subject = null, $errors = array())
-    {
-        parent::__construct($message, $code, $previous);
-        $this->_setValidationErrors($errors);
-        $this->_setValidationSubject($subject);
-    }
 }
